requirements.txt
asgiref==3.6.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
colorama==0.4.6
creategitignore==0.2.1
Django==4.2.1
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-countries==7.5.1
django-debug-toolbar==4.1.0
django-filter==23.2
django-jazzmin==2.6.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
utils==1.0.1
whitenoise==6.4.0
